chore: 提交当前更改以准备最终测试

+++ est-modules/est-messaging/est-messaging-zeromq/src/main/java/ltd/idcu/est/features/messaging/zeromq/ZeroMqConnection.java
@@ -15,8 +15,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 public class ZeroMqConnection {
     
-    private static final byte SIGNATURE = 0xFF;
-    private static final byte VERSION = 0x01;
+    private static final byte SIGNATURE = (byte) 0xFF;
+    private static final byte VERSION = (byte) 0x01;
     
     private final MessagingConfig config;
     private Socket socket;
